Revert "refactor(e2e-test): separate user flow from regression tests"
This reverts commit 1a2ea812cf62ab071de8807fc52598abbb3589cb.
